feat: Add vmname
- Add vmname is -q or --qemu option is passed.
- Factorize common code into utils.rs.
