+++ src/exporters/mod.rs
@@ -2,6 +2,7 @@ pub mod prometheus;
 pub mod qemu;
 pub mod riemann;
 pub mod stdout;
+pub mod utils;
 use clap::ArgMatches;
 use std::collections::HashMap;
